Moved fight sim stuff into separate files.
